requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.1.0
bcrypt==4.3.0
Beaker==1.12.1
beautifulsoup4==4.13.4
blivet==3.12.1
#blivet-gui==2.6.0
boto3==1.38.7
botocore==1.38.7
#Brlapi==0.8.6
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
#cockpit @ file:///builddir/build/BUILD/cockpit-338-build/cockpit-338/tmp/wheel/cockpit-338-py3-none-any.whl#sha256=dffd0430af16a3846dc473fecd8953aa2aec67072dfc968696bfcd1181f82ee3
crypt_r==3.13.1
cryptography==44.0.0
#cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
deprecation==2.1.0
distro==1.9.0
#dnf==4.23.0
ecdsa==0.19.1
fastapi==0.115.12
#fedora-third-party==0.10
feedparser==6.0.11
file-magic==0.4.0
#fros==1.1
frozenlist==1.6.0
gotrue==2.12.0
greenlet==3.2.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
iso639==0.1.4
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.68
#libcomps==0.1.21
#libdnf==0.74.0
#louis==3.28.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
multidict==6.4.3
#nftables==0.1
numpy==2.2.5
olefile==0.47
packaging==24.2
pandas==2.2.3
passlib==1.7.4
Paste==3.10.1
#perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
pluggy==1.5.0
ply==3.11
postgrest==1.0.1
productmd==1.45
progressbar2==4.5.0
propcache==0.3.1
ptyprocess==0.7.0
#pwquality==1.4.5
pyasn1==0.4.8
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.11.4
pydantic_core==2.33.2
pyenchant==3.2.2
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.62
pydantic[email]
pyOpenSSL==25.0.0
pyparted==3.13.0
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-jose==3.3.0
#python-linux-procfs==0.7.3
#python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
python-utils==3.9.0
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
#pyynl @ file:///builddir/build/BUILD/kernel-6.14.5-build/kernel-6.14.5/linux-6.14.5-300.fc42.x86_64/tools/net/ynl
RapidFuzz==3.11.0
realtime==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.23.1
#rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
rsa==4.9.1
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.15.2
#selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
#sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
#setools==4.5.1
setuptools==74.1.3
sgmllib3k==1.0.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
#sos==4.8.2
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
systemd-python==235
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
validators==0.35.0
websockets==14.2
xkbregistry==0.3
yarl==1.20.0
